Kill data-flow also in the case when data-flow bypasses a call site via normalFlow.
callToReturnFlow is only invoked, when a variable that is used at a call site reaches the call site.
Example:
when the variable a reaches (forward or backward) a call site
foo(a), the callToReturn is invoked when no callee is present.

for any other variable, say b (i.e. a variable not used at foo(a)) normalFlow is invoked.